+++ src/components/Cart/CartItemContainer/CartItemController/CartItemController.tsx
@@ -1,10 +1,10 @@
 /** @jsxImportSource @emotion/react */
 import { useRecoilState } from "recoil";
-import Checkbox from "../../../../Button/Checkbox/Checkbox";
-import { ControlStyle } from "./CartItemControls.style";
-import { checkAllItemState } from "../../../../../store/selector/selectors";
+import { ControllerStyle } from "./CartItemController.style";
+import { checkAllItemState } from "../../../../store/selector/selectors";
+import Checkbox from "../../../Button/Checkbox/Checkbox";
 
-const CartItemControls = () => {
+const CartItemController = () => {
   const [state, setState] = useRecoilState(checkAllItemState);
 
   const handleClick = () => {
@@ -14,11 +14,11 @@ const CartItemControls = () => {
   };
 
   return (
-    <div css={ControlStyle}>
+    <div css={ControllerStyle}>
       <Checkbox isCheck={state} onClick={handleClick} />
       <div>전체 선택</div>
     </div>
   );
 };
 
-export default CartItemControls;
+export default CartItemController;
