+++ src/components/Cart/CartResults/CartResults.tsx
@@ -1,12 +1,12 @@
 /** @jsxImportSource @emotion/react */
 import { IoIosInformationCircleOutline } from "react-icons/io";
 import { CartResultGuideStyle, CartResultGuideContainerStyle, CartResultsContainerStyle } from "./CartResults.style";
-import PaymentDetail from "../../../PaymentDetail/PaymentDetail";
-import Divider from "../../../Divider/Divider";
 import { useRecoilValue } from "recoil";
-import { orderAmountState, totalAmountState } from "../../../../store/selector/selectors";
-import { SHIPPING_CONSTANT } from "../../../../constants";
-import { itemIdsState } from "../../../../store/atom/atoms";
+import { orderAmountState, totalAmountState } from "../../../store/selector/selectors";
+import { itemIdsState } from "../../../store/atom/atoms";
+import { SHIPPING_CONSTANT } from "../../../constants";
+import Divider from "../../Divider/Divider";
+import PaymentDetail from "../../PaymentDetail/PaymentDetail";
 
 const CartResults = () => {
   const orderAmount = useRecoilValue(orderAmountState);
