+++ src/pages/OrderConfirmPage.tsx
@@ -1,6 +1,6 @@
 import OrderConfirmFooter from "../components/Footer/OrderConfirmFooter/OrderConfirmFooter";
 import Header from "../components/Header/Header";
-import OrderContainer from "../components/Main/Order/OrderContainer";
+import OrderContainer from "../components/OrderConfirm/OrderContainer";
 
 const OrderConfirmPage = () => {
   return (
